requirements.txt
binance-connector
typer
python-dotenv
